XIVY-7671 Use lambda and single url param test.
